refactor: move a service
